TlsRequest: Session cache; X509 chain copying
Sessions are now cached for 10 minutes, to avoid repeated handshakes on frequent connections.

There was a problem with duplicating X509 chains: while the stack was duplicated, the contained X509 certs were not. This made it possible that certs in a copied stack had been released when they were needed later.